create common component shareLink
