+++ src/features/common/components/ShareLinks/index.js
@@ -1,0 +1,1 @@
+export { ShareLinks } from './ShareLinks';
