--- src/features/donations/components/ShareDonationLinks/index.js
@@ -1,1 +1,0 @@
-export { default } from './ShareDonationLinks';
